Update generate_labels, created generate_xml.sh
Updated the generate_labels script to include a usage function, as well as support named arguments. This script will generate the label protobuf text file given a directory of annotations and a directory to save the label.pbtxt. If no directory is provided, it will generate the protobuf in the current directory.

Created the generate_xml script. This script takes a CVAT generated XML file and a directory, and splits the CVAT generated XML file into several separate XML files. Each file represents the XML annotation of a single corresponding image. The filenames are not modified. Use -h for options.